requirements.txt
asgiref==3.6.0
certifi==2022.12.7
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.1.7
django-cors-headers==3.14.0
django-countries==7.5.1
django-elasticsearch-dsl==7.3
django-filter==23.1
django-import-export==3.2.0
django-resized==1.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
elasticsearch==7.17.9
elasticsearch-dsl==7.4.1
et-xmlfile==1.1.0
Faker==18.4.0
gunicorn==20.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
phonenumbers==8.13.6
Pillow==9.4.0
psycopg2==2.9.6
PyJWT==2.6.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2022.7.1
PyYAML==6.0
six==1.16.0
sqlparse==0.4.3
tablib==3.4.0
typing_extensions==4.5.0
tzdata==2022.7
urllib3==1.26.15
whitenoise==6.4.0
xlrd==2.0.1
xlwt==1.3.0
